django template implementation v0 (homepage)
