+++ caseview/static/js/summary.js
@@ -1,0 +1,31 @@
+this.addEventListener("load", function(){
+    loadSummaryView().then((response) => {
+        if(!response){
+            console.log("Exception trace: loadCaseView()");
+            return;
+        };
+    });
+});
+
+function loadSummaryView(){
+    return new Promise((resolve) => {
+        getCaseFromDB().then((response) => {
+            if(!response){
+                console.log("Exception trace: No caseData retrieved!");
+                resolve(false);
+            }
+    
+            if(!drawCaseSentences("colmain", response)){
+                console.log("Exception trace: drawCaseSentences!");
+                resolve(false);
+            }   
+    
+            if(!showGraphInMain("graphpreview", drawGraphJudgeDistribution(response))){
+                console.log("Exception trace: showGraphInMain()");
+                resolve(false);
+            };
+
+            resolve(true);
+        });
+    });    
+}
