--- caseview/templates/case.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-
-<html xmlns="http://www.w3.org/1999/xhtml"
-lang="en-GB" xml:lang="en-GB">
-<head>
-<meta charset="UTF-8"/>
-<title>Casmoe - Case</title>
-<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3" crossorigin="anonymous"></script>
-<script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/1.8.3/jquery.min.js"></script>
-<script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-
-<!-- <link rel="stylesheet" href="{% static './css/styles.css' %}"> -->
-<script src="{% static './js/ajax_handler.js' %}"></script>
-<script src="{% static './js/file_handler.js' %}"></script>
-<script src="{% static './js/nav.js' %}"></script>
-<script src="{% static './js/lines.js' %}"></script>
-<script src="{% static './js/caseview.js' %}"></script>
-<script src="{% static './js/classes.js' %}"></script>
-<script src="{% static './js/graphs.js' %}"></script>
-
-</head>
-<body>
-<main>
-    <div class="container" id="root">
-        <!-- Header -->
-        <header class="row text-bg-success">
-            <div class="col-8 p-2 m-auto" type="button" onclick="goToHome()" id="headertext">CASMOE - Case ASMO Editor</div>
-            <div class="d-flex justify-content-center col-4 p-2">
-                <button type="button" class="btn btn-primary px-4" id="loginbutton">Login</button>
-            </div>
-        </header>
-
-        <nav class="navbar m-2 p-2 navbar-expand-md bg-light border">
-            <div class="container-fluid">
-                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                </button>
-                <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                    <ul class="navbar-nav">
-                        <li class="btn m-2 btn-light border" type="button" id="navbarBtnOverview">Overview</li>
-                        <li class="btn m-2 btn-light border" type="button" id="navbarBtnLine" onclick="loadLineView()">Line</li>
-                        <li class="btn m-2 btn-light border" type="button" id="navbarBtnGraph" onclick="loadGraphView()">Graph</li>
-                    </ul>
-                </div>                
-            </div>
-        </nav>
-
-        <div class="row border">
-            <div class="col-8 border" id="colmain">
-            </div>
-            <div class="col-4 bg-light border" id="popoutbody">
-            </div>
-        </div>
-
-        <!-- Footer  -->
-        <footer class="row text-bg-dark">
-            <div class="p-2 m-1" id="footertext">Developed by Samuel (ft19226@bristol.ac.uk). Updated June 2023</div>
-        </footer>
-    </div>
-</main>
-
-</body>
-</html>
