+++ caseview/templates/line.html
@@ -1,0 +1,62 @@
+<!DOCTYPE html>
+{% load static %}
+
+<html xmlns="http://www.w3.org/1999/xhtml"
+lang="en-GB" xml:lang="en-GB">
+<head>
+<meta charset="UTF-8"/>
+<title>Casmoe - Case</title>
+<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz" crossorigin="anonymous"></script>
+<script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/1.8.3/jquery.min.js"></script>
+<script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+
+<link rel="stylesheet" href="{% static './css/styles.css'%}"></script>
+
+<script src="{% static './js/ajax_handler.js' %}"></script>
+<script src="{% static './js/nav.js' %}"></script>
+<script src="{% static './js/line.js' %}"></script>
+<script src="{% static './js/caseview.js' %}"></script>
+
+</head>
+<body>
+<main>
+    <div class="container" id="root">
+        <!-- Header -->
+        <header class="row p-2 text-bg-success rounded-2">
+            <div class="p-2 m-auto textlogo" type="button" onclick="goToHome()" id="headertext">CASMOE - Case ASMO Editor</div>
+        </header>
+
+        <!-- Navbar -->
+        <nav class="navbar m-auto navbar-expand-md bg-light border textlogo">
+            <div class="container-fluid">
+                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+                    <span class="navbar-toggler-icon"></span>
+                </button>
+                <div class="collapse navbar-collapse" id="navbarSupportedContent">
+                    <ul class="navbar-nav">
+                        <li class="btn btn-light border-end" type="button" id="navbarBtnOverview" onclick="goToSummary()">Overview</li>
+                        <li class="btn btn-light border-end"  type="button" id="navbarBtnLine" onclick="goToLine()">Line</li>
+                        <li class="btn btn-light border-end" type="button" id="navbarBtnGraph" onclick="goToGraph()">Graph</li>
+                    </ul>
+                </div>                
+            </div>
+        </nav>
+
+        <!-- Line Content -->
+        <div class="row border m-auto">
+            <div class="col-8 border overflow-auto textlogo" id="colmain">
+            </div>
+            <div class="col-4 bg-light border textlogo" id="popoutbody">
+            </div>
+        </div>
+
+        <!-- Footer  -->
+        <footer class="row p-2 text-bg-dark rounded-2">
+            <div class="m-auto p-1 textlogo" id="footertext">Developed by Samuel (ft19226@bristol.ac.uk). Updated June 2023</div>
+        </footer>
+    </div>
+</main>
+
+</body>
+</html>
