+++ files/static/css/styles.css
@@ -19,3 +19,11 @@
     max-height: 80vh;
 }
 
+#colmain{
+    max-height: 120vh;
+}
+
+#graphlist{
+    max-height: 20vh;
+}
+
