+++ files/static/js/nav.js
@@ -4,6 +4,26 @@ function goToHome(){
     window.location.href = "/Home/";
 }
 
-function goToCaseView(){
-    window.location.href = "/Case-View/";
+function goToEditorHome(){
+    window.location.href = "/Home/Editor/";
+}
+
+function goToSearch(){
+    window.location.href='/Search/';
+}
+
+function goToReview(){
+    window.location.href='/Review/';
+}
+
+function goToGraph(){
+    window.location.href='/Case-View/Graph/';
+}
+
+function goToLine(){
+    window.location.href='/Case-View/Line/';
+}
+
+function goToSummary(){
+    window.location.href='/Case-View/Summary/';
 }
