+++ files/templates/frame_index.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+{% load static %}
+
+<html xmlns="http://www.w3.org/1999/xhtml"
+lang="en-GB" xml:lang="en-GB">
+    <head>
+        <meta charset="UTF-8"/>
+        <title>Casmoe - Review(Editor)</title>
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
+        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz" crossorigin="anonymous"></script>
+        <script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/1.8.3/jquery.min.js"></script>
+        <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+
+        <link rel="stylesheet" href="{% static './css/styles.css' %}">
+        <script src="{% static './js/ajax_handler.js' %}"></script>
+        <script src="{% static './js/file_handler.js' %}"></script>
+        <script src="{% static './js/nav.js' %}"></script>
+    </head>
+
+    <body>
+        <main>
+            <div class="container" id="root">
+                <!-- Header -->
+                {% block header %}
+                <header class="row p-2 text-bg-success rounded-2">
+                    <div class="p-2 m-auto textlogo" type="button" onclick="goToHome()" id="headertext">Case ASMO Editor</div>
+                </header>        
+                {% endblock %}
+
+                <!-- Content -->
+                {% block content %}
+                {% endblock %}
+                
+                <!-- Footer  -->
+                {% block footer %}
+                <footer class="row p-2 text-bg-dark rounded-2">
+                    <div class="m-auto p-1 textlogo" id="footertext">Developed by Samuel (ft19226@bristol.ac.uk). Updated June 2023</div>
+                </footer>
+                {% endblock %}
+            </div>
+        </main>
+    </body>
+</html>
