+++ home/static/js/review.js
@@ -1,0 +1,20 @@
+"use strict";
+
+function clearChild(elemId){
+    var elem = document.getElementById(elemId)
+    if(!elem){
+        console.log("Exception trace: Element elem not found!");
+        return false;
+    }
+    var child = elem.lastElementChild;
+    while(child){
+        elem.removeChild(child);
+        child = elem.lastElementChild;
+    }
+
+    return true;
+}
+
+function toggleEditor(){
+    
+}
