+++ home/templates/home_EDITOR.html
@@ -1,0 +1,25 @@
+{% extends "frame_editor.html" %}
+
+{% block content %}
+<div id="row">
+    <div id="homebanner" class="row p-4 m-auto my-1 rounded-2">
+        <div class="text-center p-2">
+            <div class="p-2 text-light textlogo" id="textbanner">CASMOE - Case/ASMO Editor</div>
+        </div>
+    </div>      
+    <div class="row border">
+        <div class="d-flex justify-content-center p-5 m-auto"  id="pagebuttons">
+            <div class="col-5 p-2 m-2 bg-secondary-subtle border rounded-2">
+                <div class="btn d-flex justify-content-center p-2 m-auto textlogo" onclick="goToSearch()">-- Search: --</div>
+            </div>           
+        </div>                        
+    </div>
+    <div class="row border">
+        <div class="d-flex justify-content-center p-5 m-auto"  id="pagebuttons">
+            <div class="col-5 p-2 m-2 bg-secondary-subtle border rounded-2">
+                <div class="btn d-flex justify-content-center p-2 m-auto textlogo" onclick="goToReview()">-- Review: --</div>
+            </div>           
+        </div>                        
+    </div>
+</div>
+{% endblock %}
