+++ home/urls.py
@@ -3,6 +3,6 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('', views.frontpage, name='FrontPage'),
-    path('Save-Case/', views.save_case),
+    path('', views.home_page, name='Home'),
+    path('Home/Editor/', views.home_page_editor, name='Home-Editor'),
 ]
