+++ review/templates/review.html
@@ -57,9 +57,9 @@ lang="en-GB" xml:lang="en-GB">
                 </div>
 
                 <!-- Footer  -->
-                <footer class="row text-bg-dark">
-                    <div class="p-2 m-1" id="footertext">Developed by Samuel (ft19226@bristol.ac.uk). Updated June 2023</div>
-                </footer>                
+                <footer class="row p-2 text-bg-dark rounded-2">
+                    <div class="m-auto p-1 textlogo" id="footertext">Developed by Samuel (ft19226@bristol.ac.uk). Updated June 2023</div>
+                </footer>
             </div>
         </main>
     </body>
