+++ search/apps.py
@@ -1,5 +1,6 @@
 from django.apps import AppConfig
 
-class HomeConfig(AppConfig):
+
+class SearchConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'home'
+    name = 'search'
